babel preset env. modules handled by webpack2
