+++ .babelrc
@@ -1,12 +1,22 @@
 {
-  "presets":[
-              "es2015",
-              "react",
-              "stage-2"
-            ],
-  "plugins": [
-    "add-module-exports",
-    "transform-decorators-legacy",
-    "transform-runtime"
-  ]
+  "presets": [
+    [
+      "env",
+      {
+        "targets": { "browsers": ["last 2 android versions"] },        
+        "useBuiltIns": true,
+        "debug": true,
+        "modules": false
+      }
+    ],
+    "react",
+    "es2015"
+  ],
+   "plugins": [
+      ["transform-object-rest-spread",
+        {
+          "useBuiltIns": true
+        }
+      ]
+    ]
 }
