+++ package.json
@@ -35,9 +35,9 @@
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.13.0",
     "cross-env": "^4.0.0",
     "css-loader": "^0.25.0",
     "documentation": "^4.0.0-beta16",
