+++ src/js/components/Layout/index.js
@@ -2,7 +2,7 @@ import Grid from './Grid';
 import Column from './Column';
 import Row from './Row';
 
-export default {
+export {
     Grid,
     Column,
     Row,
